Filter dialog:
o Fix my 'implemented' option "Hide inactive to-dos". The dialog didn't even read this
setting :/
o When a filter is added, set the dialog to this filter. Usability++

BTW: I had enough with developing for a while. I feel very guilty for breaking the
build and the freeze last week. I'd better focus on translating.

svn path=/trunk/kdepim/; revision=376544