After long hours of tracking down the problem that the hierarchy of the to-do list was lost in the kolab resource, here's the fix. Every now and then korganizer's tasks would loose their parents, mostly when conflict resolution is triggered.
The problem:
When generating the XML for kmail, the task checks if it has a pointer to the parent, but the CR happens before the incidence is properly set up (which is done in Calendar::addIncidence, but that is called after CR).
So the incidence does not yet have the pointer to the parent, but only the uid of the parent.
Now, during CR, the new message for kmail is created without the parent pointer -> no <parent> tag in the messge -> the next update (which happens immediately) will have the parent lost.

The solution: 
If there's no parent pointer, check if we have at least a parent UID, and if so, use this for the parent. That way the xml is generated with the correct parent and the problem is fixed.



svn path=/branches/KDE/3.5/kdepim/; revision=443891