More API documentation: Use parenthese for methods in @see line to give them better visibility among the class references.
Added documentation for the FormatFactory class, including inline example
code.

Added an examples directory and an example which shows the usage of
the format handling classes and reference it from the API docs.
The example is compiled with the application but not installed.

Added an overview text describing the rational behind the format
design using the @defgroup doxy-tag.

svn path=/trunk/kdereview/kabcclient/; revision=462058