Forward port 466810 (tilladam):
If uploading message status in the middle of a mailcheck, instead of
partially aborting the ongoing mailcheck and queuing a new one, check if
we are in the listing phase, not the download phase, and only then
(fully) abort and start afresh. If already in the downloading phase,
simply continue.

(Hunting the headers list corruption that sometimes happens during
 online imap mailchecks. This is part of getting the mailcheck process
 to where it's more predictable, so I can decently debug it.)


svn path=/trunk/KDE/kdepim/; revision=466819