forward port: SVN commit 489798 by winterz:
Provide the ability to print the incidence's Location field to the html export.

In konsolekalendar, the incidence location will always be printed to the html.
In korganizer, the location field print is configurable.

Note that permission has been granted by the docs team to introduce
a new i18n string with this patch.


svn path=/trunk/KDE/kdepim/; revision=491390