Avoid errorneous conflict resolution triggered by the following scenario:
pre-condition: 
two instances of the same event with two attendees in a folder which both 
have access to and which they both use to store the event. This leads to
two events with the same uid, one of which has an internal uid, which is
used as the uid inside korganizer, to discern the duplicates. Let alpha
be the original event, and beta the copy, with the internal uid.

Here's what should happen:

1) user A changes beta and syncs it up
2) user B syncs, get's removal of old version of beta, which triggers
   removal of the internal uid from the uid map
3) user B gets addition of the new version, with internal uid, add the
   event again, with internal uid

But due to the bug this happens:

1) as before
2) user B syncs, but KMail signals deletion of event using uid, not
   internal uid, thus internal uid is never removed from uid map
3) addition of new version of beta (with internal uid) triggers
   conflict resolution, since internal uid is still in the map

Fix: parse kolab xml in KMail to find internal uid if present 
and use it already at the stage of dcop signal emission.

(Kolab Issue 1298)

svn path=/branches/kdepim/proko2/kdepim/; revision=571107