forward port SVN commit 571208 by winterz:
Approved by Till.
Put the state of the spell checker in a status box in a compose window.
Remove the status bar message "As-you-type spelling checking enabled/disabled."

This not only looks better, but also helps me as I'm trying to debug
why spell checking sometimes comes and goes for no good reason.

svn path=/trunk/KDE/kdepim/; revision=580588