A better fix, I think, for the "Duplicated year/month in month view header" BR.
This time we simply hide MonthView's month label if we are in full view mode.
In this case, the month label from the NavigatorTab essentially becomes
the label for the MonthView, and we save the screen real estate at the
same time we eliminate the title redundancy.

To make things even more consistent and to save even more screen space
we could reduce the MonthView label from a 20pt to a 10pt font and remove
the bold-ness.  Not sure if we want to do that or not.

CCBUGS: 124066


svn path=/branches/KDE/3.5/kdepim/; revision=584857