merge SVN commit 633860 by adridg:
When the message is set to a composer with a specific cursor position,
(ie. not 0, eg. from a template containing %CURSOR) then don't reset the
cursor position when appending the signature. This makes it possible
to use %CURSOR to place the cursor anywhere in a template. Useful
for reply templates, at any rate.

svn path=/branches/KDE/3.5/kdepim/; revision=639750