Merge SVN commit 516819 by tilladam from proko2:
Change the default for hiding the groupware folders to false.


Merge SVN commit 516820 by tilladam from proko2:

Rely on the globalsettings instead of the actual runtime state of the
kmailicaliface to determine whether to filter non-groupware folders,
since during the first sync it is still turned off, but we still want to
filter already.

svn path=/branches/kdepim/enterprise/kdepim/; revision=642161